Create token.hpp
